# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical test",
        "description": "Technical test documentation",
        "contact": {
            "email": "ala@sloth-lab.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cart/delete/{id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Removes item from cart",
                "description": "Removes item from card and creating a record in Records table if it's the first time the item is deleted or increments the occurence number",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "item not found"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}